+++ pwlf/__init__.py
@@ -1,4 +1,9 @@
 from .pwlf import PiecewiseLinFit
+try:
+    import tensorflow as tf
+    from .pwlftf import PiecewiseLinFitTF
+except ImportError:
+    print('Warning: Install tensorflow to have access to PiecewiseLinFitTF.')
 import os
 
 # add rudimentary version tracking