+++ pwlf/pwlftf.py
@@ -23,23 +23,22 @@
 
 from __future__ import print_function
 # import libraries
+import tensorflow as tf
 import numpy as np
 from scipy.optimize import differential_evolution
 from scipy.optimize import fmin_l_bfgs_b
-from scipy import linalg
-from scipy import stats
 from pyDOE import lhs
+from pwlf import PiecewiseLinFit
 
 # piecewise linear fit library
 
 
-class PiecewiseLinFit(object):
+class PiecewiseLinFitTF(object):
 
-    def __init__(self, x, y, disp_res=False, sorted_data=False,
-                 lapack_driver='gelsd'):
+    def __init__(self, x, y, disp_res=False, dtype='float64', fast=True):
         r"""
         An object to fit a continuous piecewise linear function
-        to data.
+        to data using TensorFlow.
 
         Initiate the library with the supplied x and y data.
         Supply the x and y data of which you'll be fitting
@@ -58,24 +57,19 @@ class PiecewiseLinFit(object):
         disp_res : bool, optional
             Whether the optimization results should be printed. Default is
             False.
-        sorted_data : bool, optional
-            Data needs to be sorted such that x[0] <= x[1] <= ... <= x[n-1].
-            This implementation takes advantage of sorted x data in order to
-            speed up the assembly of the regression matrix. A process that
-            could be repeated several thousand times. If your data is not
-            sorted, pwlf will use numpy to sort the data. Default is False.
-        lapack_driver : str, optional
-            Which LAPACK driver is used to solve the least-squares problem.
-            Default lapack_driver='gelsd'. Options are 'gelsd', 'gelsy',
-            'gelss'. For more see
-            https://docs.scipy.org/doc/scipy/reference/generated/scipy.linalg.lstsq.html
-            http://www.netlib.org/lapack/lug/node27.html
+        dtype : str, optional
+            Data type to use for tensorflow, either 'float64' or 'float32'.
+            Default is 'float64'.
+        fast : bool, optional
+            If fast is True, then the solution is computed by solving the
+            normal equations using Cholesky decomposition. Default fast=True.
+            https://www.tensorflow.org/api_docs/python/tf/linalg/lstsq
 
         Attributes
         ----------
-        x_data : ndarray (1-D)
+        x_data : tf.tensor (n_data, 1)
             The inputted parameter x from the 1-D data set.
-        y_data : ndarray (1-D)
+        y_data : tf.tensor (n_data, 1)
             The inputted parameter y from the 1-D data set.
         n_data : int
             The number of data points.
@@ -86,8 +80,12 @@ class PiecewiseLinFit(object):
         print : bool
             Whether the optimization results should be printed. Default is
             False.
-        lapack_driver : str
-            Which LAPACK driver is used to solve the least-squares problem.
+        dtype : type
+            Data type to use for tensorflow tensors.
+        fast : boolean, optional
+            If fast is True, then the solution is computed by solving the
+            normal equations using Cholesky decomposition. Default fast=True.
+            https://www.tensorflow.org/api_docs/python/tf/linalg/lstsq
 
         Methods
         -------
@@ -141,49 +139,35 @@ class PiecewiseLinFit(object):
         Initialize for x, y data
 
         >>> import pwlf
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
 
         Initialize for x,y data and print optimization results
 
-        >>> my_pWLF = pwlf.PiecewiseLinFit(x, y, disp_res=True)
+        >>> my_pWLF = pwlf.PiecewiseLinFitTF(x, y, disp_res=True)
 
         If your data is already sorted such that x[0] <= x[1] <= ... <= x[n-1],
         use sorted_data=True for a slight performance increase while
         initializing the object
 
-        >>> my_pWLF = pwlf.PiecewiseLinFit(x, y, sorted_data=True)
+        >>> my_pWLF = pwlf.PiecewiseLinFitTF(x, y, sorted_data=True)
         """
-
-        self.print = disp_res
-        self.lapack_driver = lapack_driver
-        # x and y should be numpy arrays
-        # if they are not convert to numpy array
-        if isinstance(x, np.ndarray) is False:
-            x = np.array(x)
-        if isinstance(y, np.ndarray) is False:
-            y = np.array(y)
-
-        self.sorted_data = sorted_data
-
-        # it is assumed by default that initial arrays are not sorted
-        # i.e. if your data is already ordered
-        # from x[0] <= x[1] <= ... <= x[n-1] use sorted_data=True
-        if self.sorted_data:
-            self.x_data = x
-            self.y_data = y
+        self.fast = fast
+        if dtype == 'float64':
+            self.dtype = tf.float64
         else:
-            # sort the data from least x to max x
-            order_arg = np.argsort(x)
-            self.x_data = x[order_arg]
-            self.y_data = y[order_arg]
+            self.dtype = tf.float32
+        self.print = disp_res
         # calculate the number of data points
-        self.n_data = len(x)
+        self.n_data = x.size
 
         # set the first and last break x values to be the min and max of x
-        self.break_0 = np.min(self.x_data)
-        self.break_n = np.max(self.x_data)
+        self.break_0 = np.min(x)
+        self.break_n = np.max(x)
 
-    def assemble_regression_matrix(self, breaks, x, x_ordered=True):
+        self.x_data = tf.convert_to_tensor(x.reshape(-1, 1), dtype=self.dtype)
+        self.y_data = tf.convert_to_tensor(y.reshape(-1, 1), dtype=self.dtype)
+
+    def assemble_regression_matrix(self, breaks, x):
         r"""
         Assemble the linear regression matrix A
 
@@ -193,13 +177,9 @@ class PiecewiseLinFit(object):
             The x locations where each line segment terminates. These are
             referred to as breakpoints for each line segment. This should be
             structured as a 1-D numpy array.
-        x : ndarray (1-D)
+        x : tf.tensor (n_data, 1)
             The x locations which the linear regression matrix is assembled on.
             This must be a numpy array!
-        x_ordered : bool, optional
-            Whether x is ordered from smallest to largest. Data needs to be
-            sorted such that x[0] <= x[1] <= ... <= x[n-1].
-            Default x_ordered=False.
 
         Attributes
         ----------
@@ -213,7 +193,7 @@ class PiecewiseLinFit(object):
 
         Returns
         -------
-        A : ndarray (2-D)
+        A : tf.tensor (2-D)
             The assembled linear regression matrix.
 
         Examples
@@ -222,7 +202,7 @@ class PiecewiseLinFit(object):
         breakpoints.
 
         >>> import pwlf
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = [0.0, 0.5, 1.0]
         >>> A = assemble_regression_matrix(breaks, self.x_data)
 
@@ -230,10 +210,6 @@ class PiecewiseLinFit(object):
         # Check if breaks in ndarray, if not convert to np.array
         if isinstance(breaks, np.ndarray) is False:
             breaks = np.array(breaks)
-        if x_ordered is False:
-            # sort the data from least x to max x
-            order_arg = np.argsort(x)
-            x = x[order_arg]
 
         # Sort the breaks, then store them
         breaks_order = np.argsort(breaks)
@@ -242,23 +218,15 @@ class PiecewiseLinFit(object):
         self.n_parameters = len(breaks)
         self.n_segments = self.n_parameters - 1
 
-        # # initialize the regression matrix as zeros
-        A = np.zeros((len(x), self.n_parameters))
-        # The first two columns of the matrix are always defined as
-        A[:, 0] = 1.0
-        A[:, 1] = x - self.fit_breaks[0]
-        # Loop through the rest of A to determine the other columns
-        for i in range(self.n_segments-1):
-            # find the locations where x > breakpoint values
-            int_locations = x > self.fit_breaks[i+1]
-            if sum(int_locations) > 0:
-                # this if statement just ensures that there is at least
-                # one data point in x_c > breaks[i+1]
-                # find the first index of x where it is greater than the break
-                # point value
-                int_index = np.argmax(int_locations)
-                # only change the non-zero values of A
-                A[int_index:, i+2] = x[int_index:] - self.fit_breaks[i+1]
+        # Assemble the regression matrix
+        A_list = []
+        A_list.append(tf.ones_like(x))
+        A_list.append(x - self.fit_breaks[0])
+        zeros = tf.zeros_like(x)
+        for i in range(self.n_segments - 1):
+            mask = tf.greater(x, self.fit_breaks[i+1])
+            A_list.append(tf.where(mask, x - self.fit_breaks[i+1], zeros))
+        A = tf.concat(A_list, axis=1)
         return A
 
     def fit_with_breaks(self, breaks):
@@ -295,6 +263,8 @@ class PiecewiseLinFit(object):
             The number of line segments.
         beta : ndarray (1-D)
             The model parameters for the continuous piecewise linear fit.
+        betaTF : tf.tensor (n_parameters, 1)
+            The model parameters for the continuous piecewise linear fit.
         slopes : ndarray (1-D)
             The slope of each ling segment as a 1-D numpy array. This assumes
             that x[0] <= x[1] <= ... <= x[n]. Thus, slopes[0] is the slope
@@ -306,11 +276,6 @@ class PiecewiseLinFit(object):
         -------
         ssr : float
             Returns the sum of squares of the residuals.
-
-        Raises
-        ------
-        LinAlgError
-            This typically means your regression problem is ill-conditioned.
 
         Notes
         -----
@@ -326,7 +291,7 @@ class PiecewiseLinFit(object):
         >>> import pwlf
         >>> x = np.linspace(0.0, 1.0, 10)
         >>> y = np.random.random(10)
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = [0.0, 0.3, 0.6, 1.0]
         >>> ssr = my_pwlf.fit_with_breaks(breaks)
 
@@ -338,44 +303,21 @@ class PiecewiseLinFit(object):
 
         A = self.assemble_regression_matrix(breaks, self.x_data)
 
-        # try to solve the regression problem
-        try:
-            # least squares solver
-            beta, ssr, rank, s = linalg.lstsq(A, self.y_data,
-                                              lapack_driver=self.lapack_driver)
-
+        # least squares solver
+        beta = tf.linalg.lstsq(A, self.y_data, fast=self.fast)
+        y_hat = tf.matmul(A, beta)
+        e = y_hat - self.y_data
+        # compute the sum of square of the residuals
+        SSr = tf.matmul(tf.transpose(e), e)
+        self.betaTF = beta
+        with tf.Session():
             # save the beta parameters
-            self.beta = beta
+            self.beta = beta.eval().flatten()
+            ssr = SSr.eval()
 
-            # save the slopes
-            self.calc_slopes()
-
-            # ssr is only calculated if self.n_data > self.n_parameters
-            # in this case I'll need to calculate ssr manually
-            # where ssr = sum of square of residuals
-            if self.n_data <= self.n_parameters:
-                y_hat = np.dot(A, beta)
-                e = y_hat - self.y_data
-                ssr = np.dot(e, e)
-            if type(ssr) == list:
-                ssr = ssr[0]
-            elif type(ssr) == np.ndarray:
-                if ssr.size == 0:
-                    y_hat = np.dot(A, beta)
-                    e = y_hat - self.y_data
-                    ssr = np.dot(e, e)
-                else:
-                    ssr = ssr[0]
-
-        except linalg.LinAlgError:
-            # the computation could not converge!
-            # on an error, return ssr = np.print_function
-            # You might have a singular Matrix!!!
-            ssr = np.inf
-        if ssr is None:
-            ssr = np.inf
-            # something went wrong...
-        return ssr
+        # save the slopes
+        self.calc_slopes()
+        return ssr[0, 0]
 
     def fit_with_breaks_force_points(self, breaks, x_c, y_c):
         r"""
@@ -413,7 +355,11 @@ class PiecewiseLinFit(object):
             The number of line segments.
         beta : ndarray (1-D)
             The model parameters for the continuous piecewise linear fit.
+        betaTF : tf.tensor (n_parameters, 1)
+            The model parameters for the continuous piecewise linear fit.
         zeta : ndarray (1-D)
+            The model parameters associated with the constraint function.
+        zetaTF : tf.tensor (c_n, 1)
             The model parameters associated with the constraint function.
         slopes : ndarray (1-D)
             The slope of each ling segment as a 1-D numpy array. This assumes
@@ -421,10 +367,10 @@ class PiecewiseLinFit(object):
             of the first line segment.
         intercepts : ndarray (1-D)
             The y-intercept of each line segment as a 1-D numpy array.
-        x_c : ndarray (1-D)
+        x_c : tf.tensor (c_n, 1)
             The x locations of the data points that the piecewise linear
             function will be forced to go through.
-        y_c : ndarray (1-D)
+        y_c : tf.tensor (c_n, 1)
             The x locations of the data points that the piecewise linear
             function will be forced to go through.
         c_n : int
@@ -459,23 +405,22 @@ class PiecewiseLinFit(object):
         >>> y = np.random.random(10)
         >>> x_c = [0.0]
         >>> y_c = [0.0]
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = [0.0, 0.3, 0.6, 1.0]
         >>> L = my_pwlf.fit_with_breaks_force_points(breaks, x_c, y_c)
 
         """
-
-        # check if x_c and y_c are numpy array, if not convert to numpy array
+        self.c_n = len(x_c)
         if isinstance(x_c, np.ndarray) is False:
             x_c = np.array(x_c)
         if isinstance(y_c, np.ndarray) is False:
             y_c = np.array(y_c)
         # sort the x_c and y_c data points, then store them
         x_c_order = np.argsort(x_c)
-        self.x_c = x_c[x_c_order]
-        self.y_c = y_c[x_c_order]
-        # store the number of constraints
-        self.c_n = len(self.x_c)
+        self.x_c = tf.convert_to_tensor(x_c[x_c_order].reshape(-1, 1),
+                                        self.dtype)
+        self.y_c = tf.convert_to_tensor(y_c[x_c_order].reshape(-1, 1),
+                                        self.dtype)
 
         # Check if breaks in ndarray, if not convert to np.array
         if isinstance(breaks, np.ndarray) is False:
@@ -483,70 +428,55 @@ class PiecewiseLinFit(object):
 
         A = self.assemble_regression_matrix(breaks, self.x_data)
 
-        # Assemble the constraint matrix
-        C = np.zeros((self.c_n, self.n_parameters))
-        C[:, 0] = 1.0
-        C[:, 1] = self.x_c - self.fit_breaks[0]
-        # Loop through the rest of A to determine the other columns
-        for i in range(self.n_segments-1):
-            # find the locations where x > breakpoint values
-            int_locations = self.x_c > self.fit_breaks[i+1]
-            if sum(int_locations) > 0:
-                # this if statement just ensures that there is at least
-                # one data point in x_c > breaks[i+1]
-                # find the first index of x where it is greater than the break
-                # point value
-                int_ind = np.argmax(int_locations)
-                # only change the non-zero values of A
-                C[int_ind:, i+2] = self.x_c[int_ind:] - self.fit_breaks[i+1]
+        # penalty matrix
+        zeros = tf.zeros_like(self.x_c)
+        C_list = []
+        C_list.append(tf.ones_like(self.x_c))
+        C_list.append(self.x_c - self.fit_breaks[0])
+        for i in range(self.n_segments - 1):
+            mask = tf.greater(self.x_c, self.fit_breaks[i+1])
+            C_list.append(tf.where(mask, self.x_c - self.fit_breaks[i+1],
+                                   zeros))
+        C = tf.concat(C_list, axis=1)
 
-        # Assemble the square constrained least squares matrix
-        K = np.zeros((self.n_parameters + self.c_n,
-                      self.n_parameters + self.c_n))
-        K[0:self.n_parameters, 0:self.n_parameters] = 2.0 * np.dot(A.T, A)
-        K[:self.n_parameters, self.n_parameters:] = C.T
-        K[self.n_parameters:, :self.n_parameters] = C
-        # Assemble right hand side vector
-        yt = np.dot(2.0*A.T, self.y_data)
-        z = np.zeros(self.n_parameters + self.c_n)
-        z[:self.n_parameters] = yt
-        z[self.n_parameters:] = self.y_c
+        # assemble the KKT matrix K
+        A2T = 2.0 * tf.linalg.matmul(tf.transpose(A), A)
+        CT = tf.transpose(C)
+        K = tf.concat([A2T, CT], axis=1)
+        Cz = tf.concat([C, tf.zeros((self.c_n, self.c_n), tf.float64)], axis=1)
+        K = tf.concat([K, Cz], axis=0)
+        yt = 2.0 * tf.linalg.matmul(tf.transpose(A), self.y_data)
+        z = tf.concat([yt, self.y_c], axis=0)
 
-        # try to solve the regression problem
-        try:
-            # Solve the least squares problem
-            beta_prime = linalg.solve(K, z)
+        # solve the regression problem
+        beta = tf.linalg.matmul(tf.linalg.inv(K), z)
 
+        # compute the sum of square of the residuals
+        y_hat = tf.matmul(A, beta[0:self.n_parameters])
+        e = y_hat - self.y_data
+        SSr = tf.matmul(tf.transpose(e), e)
+
+        # Calculate the Lagrangian function
+        p = tf.matmul(CT, beta[self.n_parameters:])
+
+        self.beta_prime = beta
+        with tf.Session():
+            # save the beta parameters
+            beta_prime = beta.eval()
             # save the beta parameters
             self.beta = beta_prime[0:self.n_parameters]
+            self.betaTF = beta[0:self.n_parameters]
             # save the zeta parameters
             self.zeta = beta_prime[self.n_parameters:]
+            self.zetaTF = beta[self.n_parameters:]
+            ssr = SSr.eval()
+            L = ssr[0, 0] + np.sum(np.abs(p.eval()))
 
-            # save the slopes
-            self.calc_slopes()
-
-            # Calculate ssr
-            # where ssr = sum of square of residuals
-            y_hat = np.dot(A, self.beta)
-            e = y_hat - self.y_data
-            ssr = np.dot(e, e)
-
-            # Calculate the Lagrangian function
-            # c_x_y = np.dot(C, self.x_c.T) - self.y_c
-            p = np.dot(C.T, self.zeta)
-            L = np.sum(np.abs(p)) + ssr
-
-        except linalg.LinAlgError:
-            # the computation could not converge!
-            # on an error, return L = np.inf
-            # You might have a singular Matrix!!!
-            L = np.inf
-        if L is None:
-            L = np.inf
-            # something went wrong...
+        # save the slopes
+        self.calc_slopes()
         return L
 
-    def predict(self, x, sorted_data=False, beta=None, breaks=None):
+    def predict(self, x, beta=None, breaks=None):
         r"""
         Evaluate the fitted continuous piecewise linear function at untested
         points.
@@ -560,12 +490,6 @@ class PiecewiseLinFit(object):
         x : array_like
             The x locations where you want to predict the output of the fitted
             continuous piecewise linear function.
-        sorted_data : bool, optional
-            Data needs to be sorted such that x[0] <= x[1] <= ... <= x[n-1].
-            This implentation takes advantage of sorted x data in order to
-            speed up the assembly of the regression matrix. A processes that
-            could be repeated several thousand times. If your data is not
-            sorted, pwlf will use numpy to sort the data. Default is False.
         beta : none or ndarray (1-D), optional
             The model parameters for the continuous piecewise linear fit.
             Default is None.
@@ -584,6 +508,8 @@ class PiecewiseLinFit(object):
         n_segments : int
             The number of line segments.
         beta : ndarray (1-D)
+            The model parameters for the continuous piecewise linear fit.
+        betaTF : tf.tensor (n_parameters, 1)
             The model parameters for the continuous piecewise linear fit.
 
         Returns
@@ -604,20 +530,17 @@ class PiecewiseLinFit(object):
         >>> import pwlf
         >>> x = np.linspace(0.0, 1.0, 10)
         >>> y = np.random.random(10)
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = [0.0, 0.3, 0.6, 1.0]
         >>> ssr = my_pwlf.fit_with_breaks(breaks)
         >>> x_new = np.linspace(0.0, 1.0, 100)
         >>> yhat = my_pwlf.predict(x_new)
 
-        If the x data is already sorted you can add the sorted_data=True to
-        avoid sorting already sorted data.
-
-        >>> yhat = my_pwlf.predict(x_new, sorted_data=False)
-
         """
         if beta is not None and breaks is not None:
             self.beta = beta
+            self.betaTF = tf.convert_to_tensor(beta.reshape(-1, 1),
+                                               dtype=self.dtype)
             # Sort the breaks, then store them
             breaks_order = np.argsort(breaks)
             self.fit_breaks = breaks[breaks_order]
@@ -627,20 +550,13 @@ class PiecewiseLinFit(object):
         # check if x is numpy array, if not convert to numpy array
         if isinstance(x, np.ndarray) is False:
             x = np.array(x)
-
-        # it is assumed by default that initial arrays are not sorted
-        # i.e. if your data is already ordered
-        # from x[0] <= x[1] <= ... <= x[n-1] use sorted_data=True
-        if sorted_data is False:
-            # sort the data from least x to max x
-            order_arg = np.argsort(x)
-            x = x[order_arg]
-
-        A = self.assemble_regression_matrix(self.fit_breaks, x)
+        X = tf.convert_to_tensor(x.reshape(-1, 1), dtype=self.dtype)
+        A = self.assemble_regression_matrix(self.fit_breaks, X)
 
         # solve the regression problem
-        y_hat = np.dot(A, self.beta)
-        return y_hat
+        y_hat = tf.matmul(A, self.betaTF)
+        with tf.Session():
+            return y_hat.eval().flatten()
 
     def fit_with_breaks_opt(self, var):
         r"""
@@ -688,38 +604,22 @@ class PiecewiseLinFit(object):
 
         A = self.assemble_regression_matrix(breaks, self.x_data)
 
-        # try to solve the regression problem
+        # least squares solver
+        beta = tf.linalg.lstsq(A, self.y_data, fast=self.fast)
+        y_hat = tf.matmul(A, beta)
+        e = y_hat - self.y_data
+        # compute the sum of square of the residuals
+        SSr = tf.matmul(tf.transpose(e), e)
+        self.betaTF = beta
+        # save the beta parameters
         try:
-            # least squares solver
-            beta, ssr, rank, s = linalg.lstsq(A, self.y_data,
-                                              lapack_driver=self.lapack_driver)
-
-            # ssr is only calculated if self.n_data > self.n_parameters
-            # in all other cases I'll need to calculate ssr manually
-            # where ssr = sum of square of residuals
-            if self.n_data <= self.n_parameters:
-                y_hat = np.dot(A, beta)
-                e = y_hat - self.y_data
-                ssr = np.dot(e, e)
-            if type(ssr) == list:
-                ssr = ssr[0]
-            elif type(ssr) == np.ndarray:
-                if ssr.size == 0:
-                    y_hat = np.dot(A, beta)
-                    e = y_hat - self.y_data
-                    ssr = np.dot(e, e)
-                else:
-                    ssr = ssr[0]
-
-        except linalg.LinAlgError:
-            # the computation could not converge!
-            # on an error, return ssr = np.inf
-            # You might have a singular Matrix!!!
-            ssr = np.inf
-        if ssr is None:
-            ssr = np.inf
-            # something went wrong...
-        return ssr
+            self.beta = beta.eval().flatten()
+        except tf.errors.InvalidArgumentError:
+            # this means the Cholesky decomposition was not successful!
+            # the try; except must be placed around the executing code in tf!
+            return np.inf
+        ssr = SSr.eval()
+        return ssr[0, 0]
 
     def fit_force_points_opt(self, var):
         r"""
@@ -773,63 +673,53 @@ class PiecewiseLinFit(object):
 
         A = self.assemble_regression_matrix(breaks, self.x_data)
 
-        # Assemble the constraint matrix
-        C = np.zeros((self.c_n, self.n_parameters))
-        C[:, 0] = 1.0
-        C[:, 1] = self.x_c - breaks[0]
-        # Loop through the rest of A to determine the other columns
-        for i in range(self.n_segments-1):
-            # find the locations where x > breakpoint values
-            int_locations = self.x_c > breaks[i+1]
-            if sum(int_locations) > 0:
-                # this if statement just ensures that there is at least
-                # one data point in x_c > breaks[i+1]
-                # find the first index of x where it is greater than the break
-                # point value
-                int_index = np.argmax(int_locations)
-                # only change the non-zero values of A
-                C[int_index:, i+2] = self.x_c[int_index:] - breaks[i+1]
+        # penalty matrix
+        zeros = tf.zeros_like(self.x_c)
+        C_list = []
+        C_list.append(tf.ones_like(self.x_c))
+        C_list.append(self.x_c - self.fit_breaks[0])
+        for i in range(self.n_segments - 1):
+            mask = tf.greater(self.x_c, self.fit_breaks[i+1])
+            C_list.append(tf.where(mask, self.x_c - self.fit_breaks[i+1],
+                                   zeros))
+        C = tf.concat(C_list, axis=1)
 
-        # Assemble the square constrained least squares matrix
-        K = np.zeros((self.n_parameters + self.c_n,
-                      self.n_parameters + self.c_n))
-        K[0:self.n_parameters, 0:self.n_parameters] = 2.0 * np.dot(A.T, A)
-        K[:self.n_parameters, self.n_parameters:] = C.T
-        K[self.n_parameters:, :self.n_parameters] = C
-        # Assemble right hand side vector
-        yt = np.dot(2.0*A.T, self.y_data)
-        z = np.zeros(self.n_parameters + self.c_n)
-        z[:self.n_parameters] = yt
-        z[self.n_parameters:] = self.y_c
+        # assemble the KKT matrix K
+        A2T = 2.0 * tf.linalg.matmul(tf.transpose(A), A)
+        CT = tf.transpose(C)
+        K = tf.concat([A2T, CT], axis=1)
+        Cz = tf.concat([C, tf.zeros((self.c_n, self.c_n), tf.float64)], axis=1)
+        K = tf.concat([K, Cz], axis=0)
+        yt = 2.0 * tf.linalg.matmul(tf.transpose(A), self.y_data)
+        z = tf.concat([yt, self.y_c], axis=0)
 
-        # try to solve the regression problem
+        # solve the regression problem
+        beta = tf.linalg.matmul(tf.linalg.inv(K), z)
+
+        # compute the sum of square of the residuals
+        y_hat = tf.matmul(A, beta[0:self.n_parameters])
+        e = y_hat - self.y_data
+        SSr = tf.matmul(tf.transpose(e), e)
+
+        # Calculate the Lagrangian function
+        p = tf.matmul(CT, beta[self.n_parameters:])
+
+        self.beta_prime = beta
+        # save the beta parameters
         try:
-            # Solve the least squares problem
-            beta_prime = linalg.solve(K, z)
-
-            # save the beta parameters
-            self.beta = beta_prime[0:self.n_parameters]
-            # save the zeta parameters
-            self.zeta = beta_prime[self.n_parameters:]
-
-            # Calculate ssr
-            # where ssr = sum of square of residuals
-            y_hat = np.dot(A, self.beta)
-            e = y_hat - self.y_data
-            ssr = np.dot(e, e)
-
-            # Calculate the Lagrangian function
-            p = np.dot(C.T, self.zeta)
-            L = np.sum(np.abs(p)) + ssr
-
-        except linalg.LinAlgError:
-            # the computation could not converge!
-            # on an error, return L = np.inf
-            # You might have a singular Matrix!!!
-            L = np.inf
-        if L is None:
-            L = np.inf
-            # something went wrong...
+            beta_prime = beta.eval()
+        except tf.errors.InvalidArgumentError:
+            # the K matrix was not invertible!
+            # the try; except must be placed around the executing code in tf!
+            return np.inf
+        # save the beta parameters
+        self.beta = beta_prime[0:self.n_parameters]
+        self.betaTF = beta[0:self.n_parameters]
+        # save the zeta parameters
+        self.zeta = beta_prime[self.n_parameters:]
+        self.zetaTF = beta[self.n_parameters:]
+        ssr = SSr.eval()
+        L = ssr[0, 0] + np.sum(np.abs(p.eval()))
         return L
 
     def fit(self, n_segments, x_c=None, y_c=None, **kwargs):
@@ -869,7 +759,12 @@ class PiecewiseLinFit(object):
             The number of variables in the global optimization problem.
         beta : ndarray (1-D)
             The model parameters for the continuous piecewise linear fit.
+        betaTF : tf.tensor (n_parameters, 1)
+            The model parameters for the continuous piecewise linear fit.
         zeta : ndarray (1-D)
+            The model parameters associated with the constraint function,
+            if x_c and y_c is provided. Only created if x_c and y_c provided.
+        zetaTF : tf.tensor (c_n, 1)
             The model parameters associated with the constraint function,
             if x_c and y_c is provided. Only created if x_c and y_c provided.
         slopes : ndarray (1-D)
@@ -878,11 +773,11 @@ class PiecewiseLinFit(object):
             of the first line segment.
         intercepts : ndarray (1-D)
             The y-intercept of each line segment as a 1-D numpy array.
-        x_c : ndarray (1-D)
+        x_c : tf.tensor (c_n, 1)
             The x locations of the data points that the piecewise linear
             function will be forced to go through. Only created if x_c
             and y_c provided.
-        y_c : ndarray (1-D)
+        y_c : ndarray (c_n, 1)
             The x locations of the data points that the piecewise linear
             function will be forced to go through. Only created if x_c
             and y_c provided.
@@ -919,7 +814,7 @@ class PiecewiseLinFit(object):
         >>> import pwlf
         >>> x = np.linspace(0.0, 1.0, 10)
         >>> y = np.random.random(10)
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = my_pwlf.fit(3)
 
         Additionally you desired that the piecewise linear function go
@@ -957,10 +852,11 @@ class PiecewiseLinFit(object):
                 y_c = np.array(y_c)
             # sort the x_c and y_c data points, then store them
             x_c_order = np.argsort(x_c)
-            self.x_c = x_c[x_c_order]
-            self.y_c = y_c[x_c_order]
-            # store the number of constraints
-            self.c_n = len(self.x_c)
+            x_c_np = x_c[x_c_order]
+            y_c_np = y_c[x_c_order]
+            self.c_n = len(x_c)
+            self.x_c = tf.convert_to_tensor(x_c_np.reshape(-1, 1), self.dtype)
+            self.y_c = tf.convert_to_tensor(y_c_np.reshape(-1, 1), self.dtype)
             # Use a different function to minimize
             min_function = self.fit_force_points_opt
 
@@ -977,17 +873,18 @@ class PiecewiseLinFit(object):
         bounds[:, 1] = self.break_n
 
         # run the optimization
-        if len(kwargs) == 0:
-            res = differential_evolution(min_function, bounds,
-                                         strategy='best1bin', maxiter=1000,
-                                         popsize=50, tol=1e-3,
-                                         mutation=(0.5, 1), recombination=0.7,
-                                         seed=None, callback=None, disp=False,
-                                         polish=True, init='latinhypercube',
-                                         atol=1e-4)
-        else:
-            res = differential_evolution(min_function,
-                                         bounds, **kwargs)
+        with tf.Session():
+            if len(kwargs) == 0:
+                res = differential_evolution(min_function, bounds,
+                                             strategy='best1bin', maxiter=1000,
+                                             popsize=50, tol=1e-3,
+                                             mutation=(0.5, 1),
+                                             recombination=0.7, seed=None,
+                                             callback=None, disp=False,
+                                             polish=True,
+                                             init='latinhypercube', atol=1e-4)
+            else:
+                res = differential_evolution(min_function, bounds, **kwargs)
         if self.print is True:
             print(res)
 
@@ -1004,7 +901,7 @@ class PiecewiseLinFit(object):
         if x_c is None and y_c is None:
             self.fit_with_breaks(breaks)
         else:
-            self.fit_with_breaks_force_points(breaks, self.x_c, self.y_c)
+            self.fit_with_breaks_force_points(breaks, x_c_np, y_c_np)
 
         return self.fit_breaks
 
@@ -1052,6 +949,8 @@ class PiecewiseLinFit(object):
             The number of variables in the global optimization problem.
         beta : ndarray (1-D)
             The model parameters for the continuous piecewise linear fit.
+        betaTF : tf.tensor (n_parameters, 1)
+            The model parameters for the continuous piecewise linear fit.
         slopes : ndarray (1-D)
             The slope of each ling segment as a 1-D numpy array. This assumes
             that x[0] <= x[1] <= ... <= x[n]. Thus, slopes[0] is the slope
@@ -1087,7 +986,7 @@ class PiecewiseLinFit(object):
         >>> import pwlf
         >>> x = np.linspace(0.0, 1.0, 10)
         >>> y = np.random.random(10)
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = my_pwlf.fitfast(3)
 
         You can change the number of latin hypercube samples (or starting
@@ -1118,43 +1017,45 @@ class PiecewiseLinFit(object):
         x = np.zeros((pop, self.nVar))
         f = np.zeros(pop)
         d = []
+        with tf.Session():
+            for i, x0 in enumerate(mypop):
+                if len(kwargs) == 0:
+                    resx, resf, resd = fmin_l_bfgs_b(self.fit_with_breaks_opt,
+                                                     x0, fprime=None, args=(),
+                                                     approx_grad=True,
+                                                     bounds=bounds, m=10,
+                                                     factr=1e2, pgtol=1e-05,
+                                                     epsilon=1e-08,
+                                                     iprint=-1, maxfun=15000,
+                                                     maxiter=15000, disp=None,
+                                                     callback=None)
+                else:
+                    resx, resf, resd = fmin_l_bfgs_b(self.fit_with_breaks_opt,
+                                                     x0, fprime=None,
+                                                     approx_grad=True,
+                                                     bounds=bounds, **kwargs)
+                x[i, :] = resx
+                f[i] = resf
+                d.append(resd)
+                if self.print is True:
+                    print(i + 1, 'of ' + str(pop) + ' complete')
 
-        for i, x0 in enumerate(mypop):
-            if len(kwargs) == 0:
-                resx, resf, resd = fmin_l_bfgs_b(self.fit_with_breaks_opt, x0,
-                                                 fprime=None, args=(),
-                                                 approx_grad=True,
-                                                 bounds=bounds, m=10,
-                                                 factr=1e2, pgtol=1e-05,
-                                                 epsilon=1e-08, iprint=-1,
-                                                 maxfun=15000, maxiter=15000,
-                                                 disp=None, callback=None)
-            else:
-                resx, resf, resd = fmin_l_bfgs_b(self.fit_with_breaks_opt, x0,
-                                                 fprime=None, approx_grad=True,
-                                                 bounds=bounds, **kwargs)
-            x[i, :] = resx
-            f[i] = resf
-            d.append(resd)
+            # find the best result
+            best_ind = np.nanargmin(f)
+            best_val = f[best_ind]
+            best_break = x[best_ind]
+            res = (x[best_ind], f[best_ind], d[best_ind])
             if self.print is True:
-                print(i + 1, 'of ' + str(pop) + ' complete')
+                print(res)
 
-        # find the best result
-        best_ind = np.nanargmin(f)
-        best_val = f[best_ind]
-        best_break = x[best_ind]
-        res = (x[best_ind], f[best_ind], d[best_ind])
-        if self.print is True:
-            print(res)
+            self.ssr = best_val
 
-        self.ssr = best_val
-
-        # obtain the breakpoint locations from the best result
-        var = np.sort(best_break)
-        breaks = np.zeros(len(var) + 2)
-        breaks[1:-1] = var.copy()
-        breaks[0] = self.break_0
-        breaks[-1] = self.break_n
+            # obtain the breakpoint locations from the best result
+            var = np.sort(best_break)
+            breaks = np.zeros(len(var) + 2)
+            breaks[1:-1] = var.copy()
+            breaks[0] = self.break_0
+            breaks[-1] = self.break_n
 
         # assign parameters
         self.fit_with_breaks(breaks)
@@ -1193,6 +1094,8 @@ class PiecewiseLinFit(object):
             The number of variables in the global optimization problem.
         beta : ndarray (1-D)
             The model parameters for the continuous piecewise linear fit.
+        betaTF : tf.tensor (n_parameters, 1)
+            The model parameters for the continuous piecewise linear fit.
         slopes : ndarray (1-D)
             The slope of each ling segment as a 1-D numpy array. This assumes
             that x[0] <= x[1] <= ... <= x[n]. Thus, slopes[0] is the slope
@@ -1223,7 +1126,7 @@ class PiecewiseLinFit(object):
         >>> import pwlf
         >>> x = np.array([4., 5., 6., 7., 8.])
         >>> y = np.array([11., 13., 16., 28.92, 42.81])
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = my_pwlf.fit_guess([6.0])
 
         Note specifying one breakpoint will result in two line segments.
@@ -1242,22 +1145,22 @@ class PiecewiseLinFit(object):
         bounds = np.zeros([self.nVar, 2])
         bounds[:, 0] = self.break_0
         bounds[:, 1] = self.break_n
-
-        if len(kwargs) == 0:
-            resx, resf, _ = fmin_l_bfgs_b(self.fit_with_breaks_opt,
-                                          guess_breakpoints,
-                                          fprime=None, args=(),
-                                          approx_grad=True,
-                                          bounds=bounds, m=10,
-                                          factr=1e2, pgtol=1e-05,
-                                          epsilon=1e-08, iprint=-1,
-                                          maxfun=15000, maxiter=15000,
-                                          disp=None, callback=None)
-        else:
-            resx, resf, _ = fmin_l_bfgs_b(self.fit_with_breaks_opt,
-                                          guess_breakpoints,
-                                          fprime=None, approx_grad=True,
-                                          bounds=bounds, **kwargs)
+        with tf.Session():
+            if len(kwargs) == 0:
+                resx, resf, _ = fmin_l_bfgs_b(self.fit_with_breaks_opt,
+                                              guess_breakpoints,
+                                              fprime=None, args=(),
+                                              approx_grad=True,
+                                              bounds=bounds, m=10,
+                                              factr=1e2, pgtol=1e-05,
+                                              epsilon=1e-08, iprint=-1,
+                                              maxfun=15000, maxiter=15000,
+                                              disp=None, callback=None)
+            else:
+                resx, resf, _ = fmin_l_bfgs_b(self.fit_with_breaks_opt,
+                                              guess_breakpoints,
+                                              fprime=None, approx_grad=True,
+                                              bounds=bounds, **kwargs)
 
         self.ssr = resf
 
@@ -1307,10 +1210,10 @@ class PiecewiseLinFit(object):
             The number of variables in the global optimization problem.
         n_segments : int
             The number of line segments.
-        x_c : ndarray (1-D)
+        x_c : tf.tensor (c_n, 1)
             The x locations of the data points that the piecewise linear
             function will be forced to go through.
-        y_c : ndarray (1-D)
+        y_c : ndarray (c_n, 1)
             The x locations of the data points that the piecewise linear
             function will be forced to go through.
         c_n : int
@@ -1342,54 +1245,11 @@ class PiecewiseLinFit(object):
                 y_c = np.array(y_c)
             # sort the x_c and y_c data points, then store them
             x_c_order = np.argsort(x_c)
-            self.x_c = x_c[x_c_order]
-            self.y_c = y_c[x_c_order]
-            # store the number of constraints
             self.c_n = len(self.x_c)
-
-    def calc_slopes(self):
-        r"""
-        Calculate the slopes of the lines after a piecewise linear
-        function has been fitted.
-
-        This will also calculate the y-intercept from each line in the form
-        y = mx + b. The intercepts are stored at self.intercepts.
-
-        Attributes
-        ----------
-        slopes : ndarray (1-D)
-            The slope of each ling segment as a 1-D numpy array. This assumes
-            that x[0] <= x[1] <= ... <= x[n]. Thus, slopes[0] is the slope
-            of the first line segment.
-        intercepts : ndarray (1-D)
-            The y-intercept of each line segment as a 1-D numpy array.
-
-        Returns
-        -------
-        slopes : ndarray(1-D)
-            The slope of each ling segment as a 1-D numpy array. This assumes
-            that x[0] <= x[1] <= ... <= x[n]. Thus, slopes[0] is the slope
-            of the first line segment.
-
-        Examples
-        --------
-        Calculate the slopes after performing a simple fit
-
-        >>> import pwlf
-        >>> x = np.linspace(0.0, 1.0, 10)
-        >>> y = np.random.random(10)
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
-        >>> breaks = my_pwlf.fit(3)
-        >>> slopes = my_pwlf.slopes()
-
-        """
-        y_hat = self.predict(self.fit_breaks)
-        self.slopes = np.zeros(self.n_segments)
-        for i in range(self.n_segments):
-            self.slopes[i] = (y_hat[i+1]-y_hat[i]) / \
-                        (self.fit_breaks[i+1]-self.fit_breaks[i])
-        self.intercepts = y_hat[0:-1] - self.slopes*self.fit_breaks[0:-1]
-        return self.slopes
+            self.x_c = tf.convert_to_tensor(x_c[x_c_order].reshape(-1, 1),
+                                            self.dtype)
+            self.y_c = tf.convert_to_tensor(y_c[x_c_order].reshape(-1, 1),
+                                            self.dtype)
 
     def standard_errors(self):
         r"""
@@ -1434,40 +1294,31 @@ class PiecewiseLinFit(object):
         >>> import pwlf
         >>> x = np.linspace(0.0, 1.0, 10)
         >>> y = np.random.random(10)
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = my_pwlf.fitfast(3)
         >>> se = my_pwlf.standard_errors()
 
         """
         try:
-            nb = len(self.beta)
+            nb = self.beta.size
         except ValueError:
             errmsg = 'You do not have any beta parameters. You must perform' \
                      ' a fit before using standard_errors().'
             raise ValueError(errmsg)
 
-        ny = len(self.y_data)
+        ny = self.n_data
 
         A = self.assemble_regression_matrix(self.fit_breaks, self.x_data)
-
-        # try to solve for the standard errors
-        try:
-
-            y_hat = np.dot(A, self.beta)
-            e = y_hat - self.y_data
-
-            # solve for the unbiased estimate of variance
-            variance = np.dot(e, e) / (ny - nb)
-
-            self.se = np.sqrt(variance * (linalg.inv(np.dot(A.T,
-                                                            A)).diagonal()))
-
+        y_hat = tf.matmul(A, self.betaTF)
+        e = y_hat - self.y_data
+        variance = tf.matmul(tf.transpose(e), e) / (ny - nb)
+        AtAinv = tf.linalg.inv(tf.matmul(tf.transpose(A), A))
+        with tf.Session():
+            var = variance.eval()[0, 0] * AtAinv.eval().diagonal()
+            self.se = np.sqrt(var)
             return self.se
 
-        except linalg.LinAlgError:
-            raise linalg.LinAlgError('Singular matrix')
-
-    def prediction_variance(self, x, sorted_data=False):
+    def prediction_variance(self, x):
         r"""
         Calculate the prediction variance for each specified x location. The
         prediction variance is the uncertainty of the model due to the lack of
@@ -1483,12 +1334,6 @@ class PiecewiseLinFit(object):
         x : array_like
             The x locations where you want the prediction variance from the
             fitted continuous piecewise linear function.
-        sorted_data : bool, optional
-            Data needs to be sorted such that x[0] <= x[1] <= ... <= x[n-1].
-            This implentation takes advantage of sorted x data in order to
-            speed up the assembly of the regression matrix. A processes that
-            could be repeated several thousand times. If your data is not
-            sorted, pwlf will use numpy to sort the data. Default is False.
 
         Returns
         -------
@@ -1515,7 +1360,7 @@ class PiecewiseLinFit(object):
         >>> import pwlf
         >>> x = np.linspace(0.0, 1.0, 10)
         >>> y = np.random.random(10)
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = my_pwlf.fitfast(3)
         >>> x_new = np.linspace(0.0, 1.0, 100)
         >>> pre_var = my_pwlf.prediction_variance(x_new)
@@ -1524,52 +1369,35 @@ class PiecewiseLinFit(object):
 
         """
         try:
-            nb = len(self.beta)
+            nb = self.beta.size
         except ValueError:
             errmsg = 'You do not have any beta parameters. You must perform' \
                      ' a fit before using standard_errors().'
             raise ValueError(errmsg)
 
-        ny = len(self.y_data)
+        ny = self.n_data
 
         # check if x is numpy array, if not convert to numpy array
         if isinstance(x, np.ndarray) is False:
             x = np.array(x)
-
-        # it is assumed by default that initial arrays are not sorted
-        # i.e. if your data is already ordered
-        # from x[0] <= x[1] <= ... <= x[n-1] use sorted_data=True
-        if sorted_data is False:
-            # sort the data from least x to max x
-            order_arg = np.argsort(x)
-            x = x[order_arg]
+        # convert x to a tensor
+        x = tf.convert_to_tensor(x.reshape(-1, 1), dtype=self.dtype)
 
         # Regression matrix on training data
         Ad = self.assemble_regression_matrix(self.fit_breaks, self.x_data)
 
-        # try to solve for the unbiased variance estimation
-        try:
-
-            y_hat = np.dot(Ad, self.beta)
-            e = y_hat - self.y_data
-
-            # solve for the unbiased estimate of variance
-            variance = np.dot(e, e) / (ny - nb)
-
-        except linalg.LinAlgError:
-            raise linalg.LinAlgError('Singular matrix')
+        y_hat = tf.matmul(Ad, self.betaTF)
+        e = y_hat - self.y_data
+        variance = tf.matmul(tf.transpose(e), e) / (ny - nb)
 
         # Regression matrix on prediction data
         A = self.assemble_regression_matrix(self.fit_breaks, x)
+        AdtAdinv = tf.linalg.inv(tf.matmul(tf.transpose(Ad), Ad))
+        AAdtAdinvAT = tf.matmul(tf.matmul(A, AdtAdinv), tf.transpose(A))
 
-        # try to solve for the prediction variance at the x locations
-        try:
-            pre_var = variance * \
-                np.dot(np.dot(A, linalg.inv(np.dot(Ad.T, Ad))), A.T)
-            return pre_var.diagonal()
-
-        except linalg.LinAlgError:
-            raise linalg.LinAlgError('Singular matrix')
+        with tf.Session():
+            pre_var = variance.eval()[0, 0] * AAdtAdinvAT.eval().diagonal()
+            return pre_var
 
     def r_squared(self):
         r"""
@@ -1597,7 +1425,7 @@ class PiecewiseLinFit(object):
         >>> import pwlf
         >>> x = np.linspace(0.0, 1.0, 10)
         >>> y = np.random.random(10)
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
+        >>> my_pwlf = pwlf.PiecewiseLinFitTF(x, y)
         >>> breaks = my_pwlf.fitfast(3)
         >>> rsq = my_pwlf.r_squared()
 
@@ -1609,85 +1437,14 @@ class PiecewiseLinFit(object):
                      ' a fit before using standard_errors().'
             raise ValueError(errmsg)
         ssr = self.fit_with_breaks(fit_breaks)
-        ybar = np.ones(self.n_data) * np.mean(self.y_data)
+        ybar = tf.ones_like(self.y_data) * tf.reduce_mean(self.y_data)
+        # tf.reduce_mean defaults to flattening the entire tensor
         ydiff = self.y_data - ybar
-        try:
-            sst = np.dot(ydiff, ydiff)
-            rsq = 1.0 - (ssr/sst)
-            return rsq
-        except linalg.LinAlgError:
-            raise linalg.LinAlgError('Singular matrix')
+        sst = tf.matmul(tf.transpose(ydiff), ydiff)
+        with tf.Session():
+            rsq = 1.0 - (ssr/sst.eval()[0, 0])
+        return rsq
 
-    def p_values(self):
-        r"""
-        Calculate the p-values for each beta parameter.
+    calc_slopes = PiecewiseLinFit.calc_slopes
 
-        This calculates the p-values for the beta parameters under the
-        assumption that your breakpoint locations are known. Section 2.4.2 of
-        [2]_ defines how to calculate the p-value of individual parameters.
-        This is really a marginal test since each parameter is dependent upon
-        the other parameters.
-
-        These values are typically compared to some confidence level alpha for
-        significance. A 95% confidence level would have alpha = 0.05.
-
-        Returns
-        -------
-        p : ndarray (1-D)
-            p-values for each beta parameter where p-value[0] corresponds to
-            beta[0] and so forth
-
-        Raises
-        ------
-        ValueError
-            You have probably not performed a fit yet.
-
-        Notes
-        -----
-        This assumes that your breakpoint locations are exact! and does
-        not consider the uncertainty with your breakpoint locations.
-
-        See https://github.com/cjekel/piecewise_linear_fit_py/issues/14
-
-        References
-        ----------
-        .. [2] Myers RH, Montgomery DC, Anderson-Cook CM. Response surface
-            methodology . Hoboken. New Jersey: John Wiley & Sons, Inc.
-            2009;20:38-44.
-
-        Examples
-        --------
-        After performing a fit, one can calculate the p-value for each beta
-        parameter
-
-        >>> import pwlf
-        >>> x = np.linspace(0.0, 1.0, 10)
-        >>> y = np.random.random(10)
-        >>> my_pwlf = pwlf.PiecewiseLinFit(x, y)
-        >>> breaks = my_pwlf.fitfast(3)
-        >>> x_new = np.linspace(0.0, 1.0, 100)
-        >>> p = my_pwlf.p_values(x_new)
-
-        see also examples/standard_errrors_and_p-values.py
-
-        """
-        # calculate the standard errors associated with each beta parameter
-        # not that these standard errors and p-values are only meaningful if
-        # you have specified the specific line segment end locations
-        # at least for now...
-        self.standard_errors()
-
-        # calculate my t-value
-        t = self.beta / self.se
-
-        # degrees of freedom for t-distribution
-        n = self.n_data
-        try:
-            k = len(self.beta) - 1
-        except ValueError:
-            errmsg = 'You do not have any beta parameters. You must perform' \
-                     ' a fit before using standard_errors().'
-            raise ValueError(errmsg)
-        # calculate the p-values
-        p = 2.0 * stats.t.sf(np.abs(t), df=n-k-1)
-        return p
+    p_values = PiecewiseLinFit.p_values